♻️ refactor(websocket) [#10.9.20]: 4차 개선 - 폴링 설정 로직 완전 분리 및 문서화
🔧 변경 사항:
- **[Config]**: [web_ui/src/config/monitoring.ts](web_ui/src/config/monitoring.ts) 신규 생성
  - UI 컴포넌트에서 설정/검증 로직을 완전히 분리 (Separation of Concerns)
  - [getMetricsPollInterval](web_ui/src/config/monitoring.ts) 및 상수(DEFAULT, MIN, MAX) 중앙 관리
  - 순환 의존성 방지 및 테스트 용이성 극대화
- **[Frontend]**: [websocket-monitor.tsx](web_ui/src/components/dashboard/websocket-monitor.tsx) 경량화
  - 내부 로직 제거 후 config 모듈 import 방식으로 변경
- **[Docs]**: README_phase1.md 코드 스니펫 동기화
  - 실제 파일 구조(Config vs Component)에 맞춰 섹션 분리
  - 소스 코드의 JSDoc 주석까지 100% 일치시켜 문서 정합성 확보

🎯 목적:
- 코드 리뷰 피드백 완벽 반영 (모듈 분리 + 문서 정합성)
- 유지보수성 향상 및 아키텍처 개선
- Phase 1 WebSocket 모니터링 기능 구현 최종 마무리

✅ 검증 완료:
- 파일 분리 및 Import 경로 확인
- README 문서와 실제 코드의 일치 여부 확인

📌 Related:
- Issue [#273]
- @ai-bot-review (https://github.com/jjaayy2222/flownote-mvp/pull/377#pullrequestreview-3686084838)
